testing removing suggestions by index

+++ cogs/suggestions.py
@@ -1,7 +1,7 @@
 from discord.ext import commands
 import sqlite3
 from cogs.utils.db import add_user_to_db_or_pass
-from cogs.utils.messages import update_banner
+from cogs.utils.messages import update_banner, get_suggestion_name
 from cogs.utils.misc import check_admin_rights
 from contextlib import closing
 
@@ -54,31 +54,35 @@ class SuggestionsCog:
         else:
             user_id = ctx.author.id
             username = str(ctx.author.name)
-            game = ' '.join(data.split())
-            with closing(sqlite3.connect(self.bot.db_name)) as con:
-                with con:
-                    con.execute('UPDATE OR IGNORE Users SET username=? WHERE user_id=?;', (username, user_id))
-                    con.execute('INSERT OR IGNORE INTO Users(username, user_id) VALUES(?, ?);', (username, user_id))
-                    exists = con.execute('SELECT * FROM Suggestions \
-                                          WHERE user_id=? AND suggestion LIKE ? AND suggestion_type=?;',
-                                         (user_id, game, 'game')).fetchall()
-                if exists:
+            game = await get_suggestion_name(ctx, 'game', data)
+            if game:
+                with closing(sqlite3.connect(self.bot.db_name)) as con:
                     with con:
-                        con.execute(
-                            'DELETE FROM Suggestions WHERE user_id=? AND suggestion LIKE ? AND suggestion_type=?;',
-                            (user_id, game, 'game'))
+                        con.execute('UPDATE OR IGNORE Users SET username=? WHERE user_id=?;', (username, user_id))
+                        con.execute('INSERT OR IGNORE INTO Users(username, user_id) VALUES(?, ?);', (username, user_id))
                         exists = con.execute('SELECT * FROM Suggestions \
                                               WHERE user_id=? AND suggestion LIKE ? AND suggestion_type=?;',
                                              (user_id, game, 'game')).fetchall()
-                    if not exists:
-                        await update_banner(ctx, 'games')
-                        await ctx.send('Successfully deleted "{}" from {}\'s game suggestions'.format(game, username))
+                    if exists:
+                        with con:
+                            con.execute(
+                                'DELETE FROM Suggestions WHERE user_id=? AND suggestion LIKE ? AND suggestion_type=?;',
+                                (user_id, game, 'game'))
+                            exists = con.execute('SELECT * FROM Suggestions \
+                                                  WHERE user_id=? AND suggestion LIKE ? AND suggestion_type=?;',
+                                                 (user_id, game, 'game')).fetchall()
+                        if not exists:
+                            await update_banner(ctx, 'games')
+                            await ctx.send('Successfully deleted "{}" from {}\'s game suggestions'.format(game, username))
+                        else:
+                            await ctx.send(
+                                'Couldn\'t delete "{}" from {}\'s game suggestions, please contact Euqimor for troubleshooting'.format(
+                                    game, username))
                     else:
-                        await ctx.send(
-                            'Couldn\'t delete "{}" from {}\'s game suggestions, please contact Euqimor for troubleshooting'.format(
-                                game, username))
-                else:
-                    await ctx.send('"{}" not found in {}\'s game suggestions'.format(game, username))
+                        await ctx.send('"{}" not found in {}\'s game suggestions'.format(game, username))
+            else:
+                await ctx.send(f'Couldn\'t find anything matching "{data}" in {username}\'s suggestions')
+
 
     @commands.command(aliases=['movie_remove'])
     async def remove_movie(self, ctx, *, data):
@@ -88,81 +92,90 @@ class SuggestionsCog:
         else:
             user_id = ctx.author.id
             username = str(ctx.author.name)
-            movie = ' '.join(data.split())
-            with closing(sqlite3.connect(self.bot.db_name)) as con:
-                with con:
-                    con.execute('UPDATE OR IGNORE Users SET username=? WHERE user_id=?;', (username, user_id))
-                    con.execute('INSERT OR IGNORE INTO Users(username, user_id) VALUES(?, ?);', (username, user_id))
-                    exists = con.execute('SELECT * FROM Suggestions \
-                                          WHERE user_id=? AND suggestion LIKE ? AND suggestion_type=?;',
-                                         (user_id, movie, 'movie')).fetchall()
-                if exists:
+            movie = await get_suggestion_name(ctx, 'movie', data)
+            if movie:
+                with closing(sqlite3.connect(self.bot.db_name)) as con:
                     with con:
-                        con.execute(
-                            'DELETE FROM Suggestions WHERE user_id=? AND suggestion LIKE ? AND suggestion_type=?;',
-                            (user_id, movie, 'movie'))
+                        con.execute('UPDATE OR IGNORE Users SET username=? WHERE user_id=?;', (username, user_id))
+                        con.execute('INSERT OR IGNORE INTO Users(username, user_id) VALUES(?, ?);', (username, user_id))
                         exists = con.execute('SELECT * FROM Suggestions \
                                               WHERE user_id=? AND suggestion LIKE ? AND suggestion_type=?;',
                                              (user_id, movie, 'movie')).fetchall()
-                    if not exists:
-                        await update_banner(ctx, 'movies')
-                        await ctx.send('Successfully deleted "{}" from {}\'s movie suggestions'.format(movie, username))
+                    if exists:
+                        with con:
+                            con.execute(
+                                'DELETE FROM Suggestions WHERE user_id=? AND suggestion LIKE ? AND suggestion_type=?;',
+                                (user_id, movie, 'movie'))
+                            exists = con.execute('SELECT * FROM Suggestions \
+                                                  WHERE user_id=? AND suggestion LIKE ? AND suggestion_type=?;',
+                                                 (user_id, movie, 'movie')).fetchall()
+                        if not exists:
+                            await update_banner(ctx, 'movies')
+                            await ctx.send('Successfully deleted "{}" from {}\'s movie suggestions'.format(movie, username))
+                        else:
+                            await ctx.send(
+                                'Couldn\'t delete "{}" from {}\'s movie suggestions, please contact Euqimor for troubleshooting'
+                                .format(movie, username))
                     else:
-                        await ctx.send(
-                            'Couldn\'t delete "{}" from {}\'s movie suggestions, please contact Euqimor for troubleshooting'
-                            .format(movie, username))
-                else:
-                    await ctx.send('"{}" not found in {}\'s movie suggestions'.format(movie, username))
+                        await ctx.send('"{}" not found in {}\'s movie suggestions'.format(movie, username))
+            else:
+                await ctx.send(f'Couldn\'t find anything matching "{data}" in {username}\'s suggestions')
 
     async def adminremove(self, ctx, data):
         """Removes the game from any user's suggestions"""
-        game = ' '.join(data.split())
-        with closing(sqlite3.connect(self.bot.db_name)) as con:
-            with con:
-                exists = con.execute('SELECT * FROM Suggestions \
-                                              WHERE suggestion LIKE ? AND suggestion_type=?;',
-                                     (game, 'game')).fetchall()
-            if exists:
+        game = await get_suggestion_name(ctx, 'game', data)
+        if game:
+            with closing(sqlite3.connect(self.bot.db_name)) as con:
                 with con:
-                    con.execute('DELETE FROM Suggestions WHERE suggestion LIKE ? AND suggestion_type=?;',
-                                (game, 'game'))
                     exists = con.execute('SELECT * FROM Suggestions \
                                                   WHERE suggestion LIKE ? AND suggestion_type=?;',
                                          (game, 'game')).fetchall()
-                if not exists:
-                    await update_banner(ctx, 'games')
-                    await ctx.send('Successfully deleted "{}" from game suggestions'.format(game))
+                if exists:
+                    with con:
+                        con.execute('DELETE FROM Suggestions WHERE suggestion LIKE ? AND suggestion_type=?;',
+                                    (game, 'game'))
+                        exists = con.execute('SELECT * FROM Suggestions \
+                                                      WHERE suggestion LIKE ? AND suggestion_type=?;',
+                                             (game, 'game')).fetchall()
+                    if not exists:
+                        await update_banner(ctx, 'games')
+                        await ctx.send('Successfully deleted "{}" from game suggestions'.format(game))
+                    else:
+                        await ctx.send(
+                            'Couldn\'t delete "{}" from game suggestions, please contact Euqimor for troubleshooting'
+                            .format(game))
                 else:
-                    await ctx.send(
-                        'Couldn\'t delete "{}" from game suggestions, please contact Euqimor for troubleshooting'
-                        .format(game))
-            else:
-                await ctx.send('"{}" not found in game suggestions'.format(game))
+                    await ctx.send('"{}" not found in game suggestions'.format(game))
+        else:
+            await ctx.send(f'Couldn\'t find anything matching "{data}" in game suggestions')
 
     async def adminremove_movie(self, ctx, data):
         """Removes the movie from any user's suggestions"""
-        movie = ' '.join(data.split())
-        with closing(sqlite3.connect(self.bot.db_name)) as con:
-            with con:
-                exists = con.execute('SELECT * FROM Suggestions \
-                                      WHERE suggestion LIKE ? AND suggestion_type=?;',
-                                     (movie, 'movie')).fetchall()
-            if exists:
+        movie = await get_suggestion_name(ctx, 'movie', data)
+        if movie:
+            with closing(sqlite3.connect(self.bot.db_name)) as con:
                 with con:
-                    con.execute('DELETE FROM Suggestions WHERE suggestion LIKE ? AND suggestion_type=?;',
-                                (movie, 'movie'))
                     exists = con.execute('SELECT * FROM Suggestions \
                                           WHERE suggestion LIKE ? AND suggestion_type=?;',
                                          (movie, 'movie')).fetchall()
-                if not exists:
-                    await update_banner(ctx, 'movies')
-                    await ctx.send('Successfully deleted "{}" from movie suggestions'.format(movie))
+                if exists:
+                    with con:
+                        con.execute('DELETE FROM Suggestions WHERE suggestion LIKE ? AND suggestion_type=?;',
+                                    (movie, 'movie'))
+                        exists = con.execute('SELECT * FROM Suggestions \
+                                              WHERE suggestion LIKE ? AND suggestion_type=?;',
+                                             (movie, 'movie')).fetchall()
+                    if not exists:
+                        await update_banner(ctx, 'movies')
+                        await ctx.send('Successfully deleted "{}" from movie suggestions'.format(movie))
+                    else:
+                        await ctx.send(
+                            'Couldn\'t delete "{}" from movie suggestions, please contact Euqimor for troubleshooting'
+                            .format(movie))
                 else:
-                    await ctx.send(
-                        'Couldn\'t delete "{}" from movie suggestions, please contact Euqimor for troubleshooting'
-                        .format(movie))
-            else:
-                await ctx.send('"{}" not found in movie suggestions'.format(movie))
+                    await ctx.send('"{}" not found in movie suggestions'.format(movie))
+        else:
+            await ctx.send(f'Couldn\'t find anything matching "{data}" in movie suggestions')
 
 
 def setup(bot):
